+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-24 22:05:32
-// Commit: fix(core): prevent double tap duplicate action in core
+// Updated at: 2026-07-24 22:21:33
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '3140';
-  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
-  final DateTime timestamp = DateTime.parse('2026-07-24T22:05:32+05:00');
+  final String commitRef = '3142';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-07-24T22:21:33+05:00');
 }